fix: fixed training for correct predictions
